Changed from require to import in the tests.
