+++ src/cli.ts
@@ -5,7 +5,7 @@ const { build } = require('gluegun')
  */
 async function run(argv) {
   // create a CLI runtime
-  const cli = build()
+  const cliInterface = build()
     .brand('replicante')
     .src(__dirname)
     .plugins('./node_modules', { matching: 'replicante-*', hidden: true })
@@ -18,7 +18,7 @@ async function run(argv) {
   // this can improve performance if they're not necessary for your project:
   // .exclude(['meta', 'strings', 'print', 'filesystem', 'semver', 'system', 'prompt', 'http', 'template', 'patching', 'package-manager'])
   // and run it
-  const toolbox = await cli.run(argv)
+  const toolbox = await cliInterface.run(argv)
 
   // send it back (for testing, mostly)
   return toolbox
